refactor(models): split data and formatter modules
Extract application config, weather dataclasses, and presentation formatter helpers into focused modules while preserving existing facade imports.

Tested: uv run ruff check src tests scripts; uv run pyright; uv run pytest -q -n 0 --tb=short

Confidence: high
