+++ src/accessiweather/models/config.py
@@ -2,831 +2,8 @@
 
 from __future__ import annotations
 
-from dataclasses import dataclass, field
+from .config_app import AppConfig
+from .config_constants import CRITICAL_SETTINGS, NON_CRITICAL_SETTINGS
+from .config_settings import AppSettings
 
-from ..sound_events import DEFAULT_MUTED_SOUND_EVENTS, normalize_known_muted_sound_events
-from .weather import Location
-
-# Critical settings needed for app initialization (load synchronously)
-# These are essential for the app to start and display basic functionality
-CRITICAL_SETTINGS: set[str] = {
-    "temperature_unit",  # Display formatting
-    "data_source",  # Weather client selection
-    "update_interval_minutes",  # Background task timing
-}
-
-# Non-critical settings (defer validation until first access)
-# These can be loaded lazily without blocking startup
-NON_CRITICAL_SETTINGS: set[str] = {
-    # Alert notification settings
-    "alert_radius_type",
-    "alert_notifications_enabled",
-    "alert_notify_extreme",
-    "alert_notify_severe",
-    "alert_notify_moderate",
-    "alert_notify_minor",
-    "alert_notify_unknown",
-    "immediate_alert_details_popups",
-    "alert_global_cooldown_minutes",
-    "alert_per_alert_cooldown_minutes",
-    "alert_escalation_cooldown_minutes",
-    "alert_freshness_window_minutes",
-    "alert_max_notifications_per_hour",
-    "alert_ignored_categories",
-    # Sound settings
-    "sound_enabled",
-    "sound_pack",
-    "muted_sound_events",
-    "show_nationwide_location",
-    # Event notifications
-    "notify_discussion_update",
-    "notify_hwo_update",
-    "notify_sps_issued",
-    "notify_severe_risk_change",
-    "notify_minutely_precipitation_start",
-    "notify_minutely_precipitation_stop",
-    "minutely_precipitation_fast_polling",
-    "precipitation_sensitivity",
-    "notify_precipitation_likelihood",
-    "precipitation_likelihood_threshold",
-    # GitHub settings
-    "github_backend_url",
-    "github_app_id",
-    "github_app_private_key",
-    "github_app_installation_id",
-    # AI explanation settings
-    "openrouter_api_key",
-    "avwx_api_key",
-    "ai_model_preference",
-    "ai_explanation_style",
-    "ai_cache_ttl",
-    "custom_system_prompt",
-    "custom_instructions",
-    # API key settings (loaded lazily via keyring)
-    "pirate_weather_api_key",
-    "visual_crossing_api_key",
-    # Display preferences
-    "round_values",
-    "enable_alerts",
-    "minimize_to_tray",
-    "minimize_on_startup",
-    "startup_enabled",
-    "auto_update_enabled",
-    "update_channel",
-    "update_check_interval_hours",
-    "trend_insights_enabled",
-    "trend_hours",
-    "show_dewpoint",
-    "show_pressure_trend",
-    "show_visibility",
-    "show_uv_index",
-    "show_seasonal_data",
-    "air_quality_enabled",
-    "pollen_enabled",
-    "offline_cache_enabled",
-    "offline_cache_max_age_minutes",
-    "weather_history_enabled",
-    "forecast_duration_days",
-    "forecast_time_reference",
-    "time_display_mode",
-    "time_format_12hour",
-    "show_timezone_suffix",
-    "alert_display_style",
-    "date_format",
-    "taskbar_icon_text_enabled",
-    "taskbar_icon_dynamic_enabled",
-    "taskbar_icon_text_format",
-    "source_priority_us",
-    "source_priority_international",
-    "auto_sources_us",
-    "auto_sources_international",
-    "auto_mode_api_budget",
-    "openmeteo_weather_model",
-    "station_selection_strategy",
-    "show_impact_summaries",
-}
-
-
-@dataclass
-class AppSettings:
-    """Application settings."""
-
-    temperature_unit: str = "both"
-    update_interval_minutes: int = 10
-    enable_alerts: bool = True
-    minimize_to_tray: bool = False
-    minimize_on_startup: bool = False
-    startup_enabled: bool = False
-    data_source: str = "auto"
-    visual_crossing_api_key: str = ""
-    pirate_weather_api_key: str = ""
-    auto_update_enabled: bool = True
-    update_channel: str = "stable"
-    update_check_interval_hours: int = 24
-    sound_enabled: bool = True
-    sound_pack: str = "default"
-    muted_sound_events: list[str] = field(default_factory=lambda: list(DEFAULT_MUTED_SOUND_EVENTS))
-    # Nationwide location visibility
-    show_nationwide_location: bool = True
-    # Event-based notifications
-    notify_discussion_update: bool = True
-    notify_hwo_update: bool = True
-    notify_sps_issued: bool = True
-    notify_severe_risk_change: bool = False
-    notify_minutely_precipitation_start: bool = True
-    notify_minutely_precipitation_stop: bool = True
-    minutely_precipitation_fast_polling: bool = False
-    # Minimum intensity level required to count as precipitation ("light", "moderate", "heavy")
-    precipitation_sensitivity: str = "light"
-    notify_precipitation_likelihood: bool = False
-    precipitation_likelihood_threshold: float = 0.5
-    github_backend_url: str = ""
-    github_app_id: str = ""
-    github_app_private_key: str = ""
-    github_app_installation_id: str = ""
-    alert_radius_type: str = "county"  # "county", "point", "zone", or "state"
-    alert_notifications_enabled: bool = True
-    alert_notify_extreme: bool = True
-    alert_notify_severe: bool = True
-    alert_notify_moderate: bool = True
-    alert_notify_minor: bool = False
-    alert_notify_unknown: bool = False
-    immediate_alert_details_popups: bool = False
-    alert_global_cooldown_minutes: int = 5
-    alert_per_alert_cooldown_minutes: int = 60
-    alert_escalation_cooldown_minutes: int = 15
-    alert_freshness_window_minutes: int = 15
-    alert_max_notifications_per_hour: int = 10
-    alert_ignored_categories: list[str] = field(default_factory=list)
-    trend_insights_enabled: bool = True
-    trend_hours: int = 24
-    show_dewpoint: bool = True
-    show_pressure_trend: bool = True
-    show_visibility: bool = True
-    show_uv_index: bool = True
-    show_seasonal_data: bool = True
-    air_quality_enabled: bool = True
-    pollen_enabled: bool = True
-    offline_cache_enabled: bool = True
-    offline_cache_max_age_minutes: int = 180
-    weather_history_enabled: bool = True
-    forecast_duration_days: int = 7
-    hourly_forecast_hours: int = 6
-    forecast_time_reference: str = "location"
-    time_display_mode: str = "local"
-    time_format_12hour: bool = True
-    show_timezone_suffix: bool = False
-    # Alert dialog display style
-    alert_display_style: str = "separate"  # "separate" | "combined"
-    # Place Add/Edit/Remove Location buttons on the location row instead of the
-    # bottom button panel.  Takes effect on app restart.
-    location_buttons_on_top: bool = False
-    # Date format preset for rendered dates
-    date_format: str = "iso"  # "iso" | "us_short" | "us_long" | "eu"
-    # Taskbar icon text options
-    taskbar_icon_text_enabled: bool = False
-    taskbar_icon_dynamic_enabled: bool = True
-    taskbar_icon_text_format: str = "{temp} {condition}"
-    # Source priority settings for smart auto mode
-    source_priority_us: list[str] = field(
-        default_factory=lambda: ["nws", "openmeteo", "visualcrossing", "pirateweather"]
-    )
-    source_priority_international: list[str] = field(
-        default_factory=lambda: ["openmeteo", "pirateweather", "visualcrossing"]
-    )
-    # Open-Meteo weather model selection
-    openmeteo_weather_model: str = "best_match"
-    # NWS station selection behavior for current conditions
-    station_selection_strategy: str = "hybrid_default"
-    # AVWX API key for international aviation weather
-    avwx_api_key: str = ""
-    # AI Explanation Settings
-    openrouter_api_key: str = ""
-    ai_model_preference: str = "openrouter/free"  # free auto-router (default)
-    ai_explanation_style: str = "standard"  # "brief", "standard", "detailed"
-    ai_cache_ttl: int = 300  # 5 minutes in seconds
-    # AI Prompt Customization Settings
-    custom_system_prompt: str | None = None  # None means use default
-    custom_instructions: str | None = None  # Appended to user prompt
-    # Priority ordering settings
-    verbosity_level: str = "standard"  # "minimal", "standard", "detailed"
-    category_order: list[str] = field(
-        default_factory=lambda: [
-            "temperature",
-            "precipitation",
-            "wind",
-            "humidity_pressure",
-            "visibility_clouds",
-            "uv_index",
-        ]
-    )
-    severe_weather_override: bool = False
-    # Startup UX guidance flags
-    onboarding_wizard_shown: bool = False
-    portable_missing_api_keys_hint_shown: bool = False
-    # Display precision
-    round_values: bool = False
-    # Impact summaries (Outdoor, Driving, Allergy) — opt-in, off by default
-    show_impact_summaries: bool = False
-    # Parallel fetch timeout for smart auto mode (seconds)
-    parallel_fetch_timeout: float = 5.0
-    # Auto mode source selection — which sources participate in auto mode
-    auto_mode_api_budget: str = "max_coverage"
-    auto_sources_us: list[str] = field(
-        default_factory=lambda: ["nws", "openmeteo", "visualcrossing", "pirateweather"]
-    )
-    auto_sources_international: list[str] = field(
-        default_factory=lambda: ["openmeteo", "pirateweather", "visualcrossing"]
-    )
-
-    @staticmethod
-    def _as_bool(value, default: bool) -> bool:
-        """Normalize common truthy/falsey representations to bool."""
-        if value is None:
-            return default
-        if isinstance(value, bool):
-            return value
-        if isinstance(value, str):
-            normalized = value.strip().lower()
-            if normalized in {"true", "1", "yes", "on"}:
-                return True
-            if normalized in {"false", "0", "no", "off"}:
-                return False
-        if isinstance(value, int | float):
-            return bool(value)
-        return default
-
-    def validate_on_access(self, setting_name: str) -> bool:
-        """
-        Validate a non-critical setting on first access.
-
-        This method performs deferred validation for settings that are not
-        critical for app startup. It validates the current value and corrects
-        it to a default if invalid.
-
-        Args:
-            setting_name: The name of the setting to validate.
-
-        Returns:
-            True if the setting is valid (or was corrected), False if
-            the setting name is unknown.
-
-        """
-        if not hasattr(self, setting_name):
-            return False
-
-        value = getattr(self, setting_name)
-
-        # Validation rules for specific settings
-        if setting_name == "ai_explanation_style":
-            valid_styles = {"brief", "standard", "detailed"}
-            if value not in valid_styles:
-                setattr(self, setting_name, "standard")
-
-        elif setting_name == "update_channel":
-            valid_channels = {"stable", "beta", "dev"}
-            if value not in valid_channels:
-                setattr(self, setting_name, "stable")
-
-        elif setting_name == "time_display_mode":
-            valid_modes = {"local", "utc", "both"}
-            if value not in valid_modes:
-                setattr(self, setting_name, "local")
-
-        elif setting_name == "forecast_time_reference":
-            valid_references = {"location", "user_local"}
-            if value not in valid_references:
-                setattr(self, setting_name, "location")
-
-        elif setting_name == "forecast_duration_days":
-            if not isinstance(value, int) or value < 3 or value > 16:
-                setattr(self, setting_name, 7)
-
-        elif setting_name == "sound_pack":
-            # Ensure sound_pack is a non-empty string
-            if not isinstance(value, str) or not value.strip():
-                setattr(self, setting_name, "default")
-
-        elif setting_name == "muted_sound_events":
-            if not isinstance(value, list):
-                setattr(self, setting_name, list(DEFAULT_MUTED_SOUND_EVENTS))
-            else:
-                normalized = normalize_known_muted_sound_events(value)
-                setattr(self, setting_name, normalized)
-
-        elif setting_name == "taskbar_icon_text_format":
-            # Ensure format string is valid
-            if not isinstance(value, str) or not value.strip():
-                setattr(self, setting_name, "{temp} {condition}")
-
-        elif setting_name in {
-            "alert_global_cooldown_minutes",
-            "alert_per_alert_cooldown_minutes",
-            "alert_escalation_cooldown_minutes",
-            "alert_freshness_window_minutes",
-        }:
-            # Ensure positive integer for cooldown settings
-            if not isinstance(value, int) or value < 0:
-                defaults = {
-                    "alert_global_cooldown_minutes": 5,
-                    "alert_per_alert_cooldown_minutes": 60,
-                    "alert_escalation_cooldown_minutes": 15,
-                    "alert_freshness_window_minutes": 15,
-                }
-                setattr(self, setting_name, defaults.get(setting_name, 5))
-
-        elif setting_name == "alert_max_notifications_per_hour":
-            # Ensure positive integer
-            if not isinstance(value, int) or value < 1:
-                setattr(self, setting_name, 10)
-
-        elif setting_name == "trend_hours":
-            # Ensure reasonable range for trend hours (1-168 hours = 1 week)
-            if not isinstance(value, int) or value < 1 or value > 168:
-                setattr(self, setting_name, 24)
-
-        elif setting_name == "ai_cache_ttl":
-            # Ensure non-negative integer for cache TTL
-            if not isinstance(value, int) or value < 0:
-                setattr(self, setting_name, 300)
-
-        elif setting_name == "update_check_interval_hours":
-            # Ensure positive integer for update interval
-            if not isinstance(value, int) or value < 1:
-                setattr(self, setting_name, 24)
-
-        elif setting_name == "offline_cache_max_age_minutes":
-            # Ensure positive integer for cache age
-            if not isinstance(value, int) or value < 1:
-                setattr(self, setting_name, 180)
-
-        elif setting_name == "precipitation_sensitivity":
-            valid_levels = {"light", "moderate", "heavy"}
-            if value not in valid_levels:
-                setattr(self, setting_name, "light")
-
-        elif setting_name == "alert_radius_type":
-            valid_types = {"county", "point", "zone", "state"}
-            if value not in valid_types:
-                setattr(self, setting_name, "county")
-
-        elif setting_name == "openmeteo_weather_model":
-            valid_models = {
-                "best_match",
-                "icon_seamless",
-                "icon_global",
-                "icon_eu",
-                "icon_d2",
-                "gfs_seamless",
-                "gfs_global",
-                "ecmwf_ifs04",
-                "meteofrance_seamless",
-                "gem_seamless",
-                "jma_seamless",
-            }
-            if value not in valid_models:
-                setattr(self, setting_name, "best_match")
-
-        elif setting_name == "station_selection_strategy":
-            valid_strategies = {
-                "nearest",
-                "major_airport_preferred",
-                "freshest_observation",
-                "hybrid_default",
-            }
-            if value not in valid_strategies:
-                setattr(self, setting_name, "hybrid_default")
-
-        elif setting_name == "auto_mode_api_budget":
-            valid_budgets = {"economy", "balanced", "max_coverage"}
-            if value not in valid_budgets:
-                setattr(self, setting_name, "max_coverage")
-
-        elif setting_name == "alert_display_style":
-            valid_styles = {"separate", "combined"}
-            if value not in valid_styles:
-                setattr(self, setting_name, "separate")
-
-        elif setting_name == "date_format":
-            valid_formats = {"iso", "us_short", "us_long", "eu"}
-            if value not in valid_formats:
-                setattr(self, setting_name, "iso")
-
-        elif setting_name in {
-            "source_priority_us",
-            "source_priority_international",
-            "auto_sources_us",
-            "auto_sources_international",
-        }:
-            # Ensure valid list of source names
-            valid_sources = {"nws", "openmeteo", "visualcrossing", "pirateweather"}
-            us_default = ["nws", "openmeteo", "visualcrossing", "pirateweather"]
-            intl_default = ["openmeteo", "pirateweather", "visualcrossing"]
-            is_us_setting = setting_name in {"source_priority_us", "auto_sources_us"}
-            if not isinstance(value, list):
-                setattr(self, setting_name, us_default if is_us_setting else intl_default)
-            else:
-                # Filter to only valid sources
-                filtered = [s for s in value if s in valid_sources]
-                if not filtered:
-                    setattr(self, setting_name, us_default if is_us_setting else intl_default)
-                elif filtered != value:
-                    setattr(self, setting_name, filtered)
-
-        elif setting_name == "alert_ignored_categories":
-            # Ensure it's a list
-            if not isinstance(value, list):
-                setattr(self, setting_name, [])
-
-        # Boolean settings are validated by _as_bool during from_dict,
-        # but we can still ensure they're actually booleans
-        elif setting_name in NON_CRITICAL_SETTINGS:
-            # Get the default value for this setting from a fresh instance
-            default_settings = AppSettings()
-            default_value = getattr(default_settings, setting_name, None)
-            if isinstance(default_value, bool) and not isinstance(value, bool):
-                setattr(self, setting_name, self._as_bool(value, default_value))
-
-        return True
-
-    def to_dict(self) -> dict:
-        """Convert to dictionary for JSON serialization."""
-        return {
-            "temperature_unit": self.temperature_unit,
-            "update_interval_minutes": self.update_interval_minutes,
-            "enable_alerts": self.enable_alerts,
-            "minimize_to_tray": self.minimize_to_tray,
-            "minimize_on_startup": self.minimize_on_startup,
-            "startup_enabled": self.startup_enabled,
-            "data_source": self.data_source,
-            # weather provider API keys and github_app_* are stored in secure keyring, not JSON
-            "auto_update_enabled": self.auto_update_enabled,
-            "update_channel": self.update_channel,
-            "update_check_interval_hours": self.update_check_interval_hours,
-            "sound_enabled": self.sound_enabled,
-            "sound_pack": self.sound_pack,
-            "muted_sound_events": self.muted_sound_events,
-            "show_nationwide_location": self.show_nationwide_location,
-            "notify_discussion_update": self.notify_discussion_update,
-            "notify_hwo_update": self.notify_hwo_update,
-            "notify_sps_issued": self.notify_sps_issued,
-            "notify_severe_risk_change": self.notify_severe_risk_change,
-            "notify_minutely_precipitation_start": self.notify_minutely_precipitation_start,
-            "notify_minutely_precipitation_stop": self.notify_minutely_precipitation_stop,
-            "minutely_precipitation_fast_polling": self.minutely_precipitation_fast_polling,
-            "precipitation_sensitivity": self.precipitation_sensitivity,
-            "notify_precipitation_likelihood": self.notify_precipitation_likelihood,
-            "precipitation_likelihood_threshold": self.precipitation_likelihood_threshold,
-            "github_backend_url": self.github_backend_url,
-            "alert_radius_type": self.alert_radius_type,
-            "alert_notifications_enabled": self.alert_notifications_enabled,
-            "alert_notify_extreme": self.alert_notify_extreme,
-            "alert_notify_severe": self.alert_notify_severe,
-            "alert_notify_moderate": self.alert_notify_moderate,
-            "alert_notify_minor": self.alert_notify_minor,
-            "alert_notify_unknown": self.alert_notify_unknown,
-            "immediate_alert_details_popups": self.immediate_alert_details_popups,
-            "alert_global_cooldown_minutes": self.alert_global_cooldown_minutes,
-            "alert_per_alert_cooldown_minutes": self.alert_per_alert_cooldown_minutes,
-            "alert_escalation_cooldown_minutes": self.alert_escalation_cooldown_minutes,
-            "alert_freshness_window_minutes": self.alert_freshness_window_minutes,
-            "alert_max_notifications_per_hour": self.alert_max_notifications_per_hour,
-            "alert_ignored_categories": self.alert_ignored_categories,
-            "trend_insights_enabled": self.trend_insights_enabled,
-            "trend_hours": self.trend_hours,
-            "show_dewpoint": self.show_dewpoint,
-            "show_pressure_trend": self.show_pressure_trend,
-            "show_visibility": self.show_visibility,
-            "show_uv_index": self.show_uv_index,
-            "show_seasonal_data": self.show_seasonal_data,
-            "air_quality_enabled": self.air_quality_enabled,
-            "pollen_enabled": self.pollen_enabled,
-            "offline_cache_enabled": self.offline_cache_enabled,
-            "offline_cache_max_age_minutes": self.offline_cache_max_age_minutes,
-            "weather_history_enabled": self.weather_history_enabled,
-            "forecast_duration_days": self.forecast_duration_days,
-            "hourly_forecast_hours": self.hourly_forecast_hours,
-            "forecast_time_reference": self.forecast_time_reference,
-            "time_display_mode": self.time_display_mode,
-            "time_format_12hour": self.time_format_12hour,
-            "show_timezone_suffix": self.show_timezone_suffix,
-            "alert_display_style": self.alert_display_style,
-            "location_buttons_on_top": self.location_buttons_on_top,
-            "date_format": self.date_format,
-            "taskbar_icon_text_enabled": self.taskbar_icon_text_enabled,
-            "taskbar_icon_dynamic_enabled": self.taskbar_icon_dynamic_enabled,
-            "taskbar_icon_text_format": self.taskbar_icon_text_format,
-            "source_priority_us": self.source_priority_us,
-            "source_priority_international": self.source_priority_international,
-            "auto_mode_api_budget": self.auto_mode_api_budget,
-            "auto_sources_us": self.auto_sources_us,
-            "auto_sources_international": self.auto_sources_international,
-            "openmeteo_weather_model": self.openmeteo_weather_model,
-            "station_selection_strategy": self.station_selection_strategy,
-            # AI settings and AVWX key stored in secure storage, not here
-            "ai_model_preference": self.ai_model_preference,
-            "ai_explanation_style": self.ai_explanation_style,
-            "ai_cache_ttl": self.ai_cache_ttl,
-            # AI Prompt Customization
-            "custom_system_prompt": self.custom_system_prompt,
-            "custom_instructions": self.custom_instructions,
-            # Priority ordering settings
-            "verbosity_level": self.verbosity_level,
-            "category_order": self.category_order,
-            "severe_weather_override": self.severe_weather_override,
-            "onboarding_wizard_shown": self.onboarding_wizard_shown,
-            "portable_missing_api_keys_hint_shown": self.portable_missing_api_keys_hint_shown,
-            "round_values": self.round_values,
-            "show_impact_summaries": self.show_impact_summaries,
-        }
-
-    @classmethod
-    def from_dict(cls, data: dict) -> AppSettings:
-        """Create from dictionary."""
-        settings = cls(
-            temperature_unit=data.get("temperature_unit", "both"),
-            update_interval_minutes=data.get("update_interval_minutes", 10),
-            enable_alerts=cls._as_bool(data.get("enable_alerts"), True),
-            minimize_to_tray=cls._as_bool(data.get("minimize_to_tray"), False),
-            minimize_on_startup=cls._as_bool(data.get("minimize_on_startup"), False),
-            startup_enabled=cls._as_bool(data.get("startup_enabled"), False),
-            data_source=data.get("data_source", "auto"),
-            visual_crossing_api_key=data.get("visual_crossing_api_key", ""),
-            pirate_weather_api_key=data.get("pirate_weather_api_key", ""),
-            auto_update_enabled=cls._as_bool(data.get("auto_update_enabled"), True),
-            update_channel=data.get("update_channel", "stable"),
-            update_check_interval_hours=data.get("update_check_interval_hours", 24),
-            sound_enabled=cls._as_bool(data.get("sound_enabled"), True),
-            sound_pack=data.get("sound_pack", "default"),
-            muted_sound_events=data.get("muted_sound_events", list(DEFAULT_MUTED_SOUND_EVENTS)),
-            show_nationwide_location=cls._as_bool(data.get("show_nationwide_location"), True),
-            notify_discussion_update=cls._as_bool(data.get("notify_discussion_update"), True),
-            notify_hwo_update=cls._as_bool(data.get("notify_hwo_update"), True),
-            notify_sps_issued=cls._as_bool(data.get("notify_sps_issued"), True),
-            notify_severe_risk_change=cls._as_bool(data.get("notify_severe_risk_change"), False),
-            notify_minutely_precipitation_start=cls._as_bool(
-                data.get("notify_minutely_precipitation_start"), True
-            ),
-            notify_minutely_precipitation_stop=cls._as_bool(
-                data.get("notify_minutely_precipitation_stop"), True
-            ),
-            minutely_precipitation_fast_polling=cls._as_bool(
-                data.get("minutely_precipitation_fast_polling"), False
-            ),
-            precipitation_sensitivity=data.get("precipitation_sensitivity", "light"),
-            notify_precipitation_likelihood=cls._as_bool(
-                data.get("notify_precipitation_likelihood"), False
-            ),
-            precipitation_likelihood_threshold=float(
-                data.get("precipitation_likelihood_threshold", 0.5)
-            ),
-            github_backend_url=data.get("github_backend_url", ""),
-            alert_radius_type=data.get("alert_radius_type", "county"),
-            alert_notifications_enabled=cls._as_bool(data.get("alert_notifications_enabled"), True),
-            alert_notify_extreme=cls._as_bool(data.get("alert_notify_extreme"), True),
-            alert_notify_severe=cls._as_bool(data.get("alert_notify_severe"), True),
-            alert_notify_moderate=cls._as_bool(data.get("alert_notify_moderate"), True),
-            alert_notify_minor=cls._as_bool(data.get("alert_notify_minor"), False),
-            alert_notify_unknown=cls._as_bool(data.get("alert_notify_unknown"), False),
-            immediate_alert_details_popups=cls._as_bool(
-                data.get("immediate_alert_details_popups"), False
-            ),
-            alert_global_cooldown_minutes=data.get("alert_global_cooldown_minutes", 5),
-            alert_per_alert_cooldown_minutes=data.get("alert_per_alert_cooldown_minutes", 60),
-            alert_escalation_cooldown_minutes=data.get("alert_escalation_cooldown_minutes", 15),
-            alert_freshness_window_minutes=data.get("alert_freshness_window_minutes", 15),
-            alert_max_notifications_per_hour=data.get("alert_max_notifications_per_hour", 10),
-            alert_ignored_categories=data.get("alert_ignored_categories", []),
-            trend_insights_enabled=cls._as_bool(data.get("trend_insights_enabled"), True),
-            trend_hours=data.get("trend_hours", 24),
-            show_dewpoint=cls._as_bool(data.get("show_dewpoint"), True),
-            show_pressure_trend=cls._as_bool(data.get("show_pressure_trend"), True),
-            show_visibility=cls._as_bool(data.get("show_visibility"), True),
-            show_uv_index=cls._as_bool(data.get("show_uv_index"), True),
-            show_seasonal_data=cls._as_bool(data.get("show_seasonal_data"), True),
-            air_quality_enabled=cls._as_bool(data.get("air_quality_enabled"), True),
-            pollen_enabled=cls._as_bool(data.get("pollen_enabled"), True),
-            offline_cache_enabled=cls._as_bool(data.get("offline_cache_enabled"), True),
-            offline_cache_max_age_minutes=data.get("offline_cache_max_age_minutes", 180),
-            weather_history_enabled=cls._as_bool(data.get("weather_history_enabled"), True),
-            forecast_duration_days=data.get("forecast_duration_days", 7),
-            hourly_forecast_hours=data.get("hourly_forecast_hours", 6),
-            forecast_time_reference=data.get("forecast_time_reference", "location"),
-            time_display_mode=data.get("time_display_mode", "local"),
-            time_format_12hour=cls._as_bool(data.get("time_format_12hour"), True),
-            show_timezone_suffix=cls._as_bool(data.get("show_timezone_suffix"), False),
-            alert_display_style=data.get("alert_display_style", "separate"),
-            location_buttons_on_top=cls._as_bool(data.get("location_buttons_on_top"), False),
-            date_format=data.get("date_format", "iso"),
-            taskbar_icon_text_enabled=cls._as_bool(data.get("taskbar_icon_text_enabled"), False),
-            taskbar_icon_dynamic_enabled=cls._as_bool(
-                data.get("taskbar_icon_dynamic_enabled"), True
-            ),
-            taskbar_icon_text_format=data.get("taskbar_icon_text_format", "{temp} {condition}"),
-            source_priority_us=data.get(
-                "source_priority_us", ["nws", "openmeteo", "visualcrossing", "pirateweather"]
-            ),
-            source_priority_international=data.get(
-                "source_priority_international", ["openmeteo", "pirateweather", "visualcrossing"]
-            ),
-            auto_mode_api_budget=data.get("auto_mode_api_budget", "max_coverage"),
-            auto_sources_us=data.get(
-                "auto_sources_us", ["nws", "openmeteo", "visualcrossing", "pirateweather"]
-            ),
-            auto_sources_international=data.get(
-                "auto_sources_international", ["openmeteo", "pirateweather", "visualcrossing"]
-            ),
-            openmeteo_weather_model=data.get("openmeteo_weather_model", "best_match"),
-            station_selection_strategy=data.get("station_selection_strategy", "hybrid_default"),
-            # AVWX and AI settings (stored in secure storage)
-            avwx_api_key=data.get("avwx_api_key", ""),
-            openrouter_api_key=data.get("openrouter_api_key", ""),
-            ai_model_preference=data.get("ai_model_preference", "openrouter/free"),
-            ai_explanation_style=data.get("ai_explanation_style", "standard"),
-            ai_cache_ttl=data.get("ai_cache_ttl", 300),
-            # AI Prompt Customization
-            custom_system_prompt=data.get("custom_system_prompt"),
-            custom_instructions=data.get("custom_instructions"),
-            # Priority ordering settings
-            verbosity_level=data.get("verbosity_level", "standard"),
-            category_order=data.get(
-                "category_order",
-                [
-                    "temperature",
-                    "precipitation",
-                    "wind",
-                    "humidity_pressure",
-                    "visibility_clouds",
-                    "uv_index",
-                ],
-            ),
-            severe_weather_override=cls._as_bool(data.get("severe_weather_override"), False),
-            onboarding_wizard_shown=cls._as_bool(data.get("onboarding_wizard_shown"), False),
-            portable_missing_api_keys_hint_shown=cls._as_bool(
-                data.get("portable_missing_api_keys_hint_shown"), False
-            ),
-            round_values=cls._as_bool(data.get("round_values"), False),
-            show_impact_summaries=cls._as_bool(data.get("show_impact_summaries"), False),
-        )
-        settings.validate_on_access("auto_mode_api_budget")
-        settings.validate_on_access("alert_display_style")
-        settings.validate_on_access("date_format")
-        return settings
-
-    def to_alert_settings(self):
-        """Convert to AlertSettings for the alert management system."""
-        from accessiweather.alert_manager import AlertSettings
-
-        settings = AlertSettings()
-        settings.notifications_enabled = self.alert_notifications_enabled
-        settings.sound_enabled = self.sound_enabled
-        settings.global_cooldown = self.alert_global_cooldown_minutes
-        settings.per_alert_cooldown = self.alert_per_alert_cooldown_minutes
-        settings.escalation_cooldown = self.alert_escalation_cooldown_minutes
-        settings.freshness_window_minutes = self.alert_freshness_window_minutes
-        settings.max_notifications_per_hour = self.alert_max_notifications_per_hour
-        settings.ignored_categories = set(self.alert_ignored_categories)
-
-        if self.alert_notify_unknown:
-            settings.min_severity_priority = 1
-        elif self.alert_notify_minor:
-            settings.min_severity_priority = 2
-        elif self.alert_notify_moderate:
-            settings.min_severity_priority = 3
-        elif self.alert_notify_severe:
-            settings.min_severity_priority = 4
-        elif self.alert_notify_extreme:
-            settings.min_severity_priority = 5
-        else:
-            settings.min_severity_priority = 6
-
-        return settings
-
-
-@dataclass
-class AppConfig:
-    """Application configuration."""
-
-    settings: AppSettings
-    locations: list[Location]
-    current_location: Location | None = None
-
-    def to_dict(self) -> dict:
-        """Convert to dictionary for JSON serialization."""
-        return {
-            "settings": self.settings.to_dict(),
-            "locations": [
-                {
-                    "name": loc.name,
-                    "latitude": loc.latitude,
-                    "longitude": loc.longitude,
-                    **({"country_code": loc.country_code} if loc.country_code else {}),
-                    **({"marine_mode": True} if loc.marine_mode else {}),
-                    **({"timezone": loc.timezone} if loc.timezone else {}),
-                    **({"forecast_zone_id": loc.forecast_zone_id} if loc.forecast_zone_id else {}),
-                    **({"cwa_office": loc.cwa_office} if loc.cwa_office else {}),
-                    **({"county_zone_id": loc.county_zone_id} if loc.county_zone_id else {}),
-                    **({"fire_zone_id": loc.fire_zone_id} if loc.fire_zone_id else {}),
-                    **({"radar_station": loc.radar_station} if loc.radar_station else {}),
-                }
-                for loc in self.locations
-            ],
-            "current_location": {
-                "name": self.current_location.name,
-                "latitude": self.current_location.latitude,
-                "longitude": self.current_location.longitude,
-                **(
-                    {"country_code": self.current_location.country_code}
-                    if self.current_location.country_code
-                    else {}
-                ),
-                **({"marine_mode": True} if self.current_location.marine_mode else {}),
-                **(
-                    {"timezone": self.current_location.timezone}
-                    if self.current_location.timezone
-                    else {}
-                ),
-                **(
-                    {"forecast_zone_id": self.current_location.forecast_zone_id}
-                    if self.current_location.forecast_zone_id
-                    else {}
-                ),
-                **(
-                    {"cwa_office": self.current_location.cwa_office}
-                    if self.current_location.cwa_office
-                    else {}
-                ),
-                **(
-                    {"county_zone_id": self.current_location.county_zone_id}
-                    if self.current_location.county_zone_id
-                    else {}
-                ),
-                **(
-                    {"fire_zone_id": self.current_location.fire_zone_id}
-                    if self.current_location.fire_zone_id
-                    else {}
-                ),
-                **(
-                    {"radar_station": self.current_location.radar_station}
-                    if self.current_location.radar_station
-                    else {}
-                ),
-            }
-            if self.current_location
-            else None,
-        }
-
-    @classmethod
-    def from_dict(cls, data: dict) -> AppConfig:
-        """Create from dictionary."""
-        settings = AppSettings.from_dict(data.get("settings", {}))
-
-        locations = []
-        for loc_data in data.get("locations", []):
-            locations.append(
-                Location(
-                    name=loc_data["name"],
-                    latitude=loc_data["latitude"],
-                    longitude=loc_data["longitude"],
-                    timezone=loc_data.get("timezone"),
-                    country_code=loc_data.get("country_code"),
-                    marine_mode=bool(loc_data.get("marine_mode", False)),
-                    forecast_zone_id=loc_data.get("forecast_zone_id"),
-                    cwa_office=loc_data.get("cwa_office"),
-                    county_zone_id=loc_data.get("county_zone_id"),
-                    fire_zone_id=loc_data.get("fire_zone_id"),
-                    radar_station=loc_data.get("radar_station"),
-                )
-            )
-
-        current_location = None
-        if data.get("current_location"):
-            loc_data = data["current_location"]
-            current_location = Location(
-                name=loc_data["name"],
-                latitude=loc_data["latitude"],
-                longitude=loc_data["longitude"],
-                timezone=loc_data.get("timezone"),
-                country_code=loc_data.get("country_code"),
-                marine_mode=bool(loc_data.get("marine_mode", False)),
-                forecast_zone_id=loc_data.get("forecast_zone_id"),
-                cwa_office=loc_data.get("cwa_office"),
-                county_zone_id=loc_data.get("county_zone_id"),
-                fire_zone_id=loc_data.get("fire_zone_id"),
-                radar_station=loc_data.get("radar_station"),
-            )
-
-        return cls(
-            settings=settings,
-            locations=locations,
-            current_location=current_location,
-        )
-
-    @classmethod
-    def default(cls) -> AppConfig:
-        """Create default configuration."""
-        return cls(
-            settings=AppSettings(),
-            locations=[],
-            current_location=None,
-        )
+__all__ = ["AppConfig", "AppSettings", "CRITICAL_SETTINGS", "NON_CRITICAL_SETTINGS"]
